// src/api/models/userResponse.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface UserResponse {
  id?: number;
  username?: string;
  email?: string;
  phone?: string;
  avatar?: string;
  status?: number;
  createdAt?: string;
  updatedAt?: string;
}